+++ school-medical-management/src/main/java/com/swp391/school_medical_management/modules/users/repositories/HealthCheckFormRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface HealthCheckFormRepository extends JpaRepository<HealthCheckFormEntity, Long> {
+public interface HealthCheckFormRepository extends JpaRepository<HealthCheckFormEntity, Integer> {
     // List<HealthCheckFormEntity> findHealthCheckFormEntityByHealthCheckProgramAndStudent(
     //                 HealthCheckProgramEntity healthCheckProgramEntity, StudentEntity student);
